+++ bundle/src/Mail/Component/Button.php
@@ -9,22 +9,10 @@ use Symfony\UX\TwigComponent\Attribute\ExposeInTemplate;
 class Button
 {
 
-    /**
-     * @var 'small' | 'medium' | 'large'
-     */
-    public string $size = 'medium';
-    public string $href = '';
+    // public string $size = 'medium';
 
-    #[ExposeInTemplate]
-    public function getStyles(): string
-    {
-        $styles = [
-            'small' => 'padding: 2px 12px; height: 26px; font-size: 13px;',
-            'medium' => 'padding: 2px 14px; height: 30px; font-size: 14px;',
-            'large' => 'padding: 2px 20px; height: 36px; font-size: 14px;',
-        ];
-        return $styles[$this->size];
-    }
+
+    public string $href = '';
 
 
 }