+++ src/Http/Middleware/AuthMiddleware.php
@@ -4,6 +4,7 @@ namespace Hyvor\Internal\Http\Middleware;
 
 use Closure;
 use Hyvor\Internal\Auth\Auth;
+use Hyvor\Internal\Auth\AuthInterface;
 use Hyvor\Internal\Http\Exceptions\HttpException;
 use Illuminate\Http\Request;
 
@@ -13,7 +14,7 @@ class AuthMiddleware
     public function handle(Request $request, Closure $next): mixed
     {
         $cookie = $request->cookie(Auth::HYVOR_SESSION_COOKIE_NAME);
-        $user = is_string($cookie) && $cookie ? app(Auth::class)->check($cookie) : null;
+        $user = app(AuthInterface::class)->check(is_string($cookie) ? $cookie : '');
 
         if (!$user) {
             throw new HttpException('Unauthorized', 401);