+++ tests/Unit/Auth/AuthFakeTestTrait.php
@@ -1,0 +1,211 @@
+<?php
+
+namespace Hyvor\Internal\Tests\Unit\Auth;
+
+use Hyvor\Internal\Auth\AuthFake;
+use Hyvor\Internal\Auth\AuthInterface;
+use Hyvor\Internal\Auth\AuthUser;
+
+trait AuthFakeTestTrait
+{
+
+    protected function getAuthFake(): AuthFake
+    {
+        $provider = $this->getContainer()->get(AuthInterface::class);
+        assert($provider instanceof AuthFake);
+        return $provider;
+    }
+
+    abstract protected function enable(?array $user = null): void;
+
+    public function testCheckBasedOnUserIdConfig(): void
+    {
+        $this->enable(['id' => 1]);
+        $this->assertSame(1, $this->getAuthFake()->user?->id);
+
+        $this->enable(['id' => 2]);
+        $this->assertSame(2, $this->getAuthFake()->user?->id);
+
+        $this->enable(null);
+        $this->assertNull($this->getAuthFake()->user);
+    }
+
+    public function testDatabaseHelperFunctions(): void
+    {
+        $this->enable();
+        AuthFake::databaseSet([
+            ['id' => 1, 'name' => 'John'],
+            ['id' => 2, 'name' => 'Jane'],
+        ]);
+
+        $db = AuthFake::databaseGet();
+        $this->assertNotNull($db);
+        $this->assertCount(2, $db);
+        $this->assertSame('John', $db[0]->name);
+        $this->assertSame('Jane', $db[1]->name);
+
+        AuthFake::databaseAdd(['id' => 3, 'name' => 'Jack']);
+        $this->assertCount(3, $db);
+        $this->assertSame('Jack', $db[2]->name);
+
+        AuthFake::databaseClear();
+        $this->assertNull(AuthFake::databaseGet());
+    }
+
+    public function test_database_add(): void
+    {
+        $this->enable();
+        AuthFake::databaseAdd(['id' => 3, 'name' => 'Jack']);
+        $db = AuthFake::databaseGet();
+        $this->assertNotNull($db);
+        $this->assertCount(1, $db);
+        $this->assertSame('Jack', $db[0]->name);
+    }
+
+    public function test_database_set_user_array(): void
+    {
+        $this->enable();
+        AuthFake::databaseSet([
+            AuthUser::fromArray([
+                'id' => 1,
+                'name' => 'Supun',
+                'email' => 'supun@hyvor.com',
+                'username' => 'supun'
+            ])
+        ]);
+        $db = AuthFake::databaseGet();
+
+        $this->assertNotNull($db);
+        $this->assertSame(1, $db[0]->id);
+        $this->assertSame('Supun', $db[0]->name);
+    }
+
+    public function testFromId(): void
+    {
+        $this->enable();
+        $id20 = $this->getAuthFake()->fromId(20);
+        $this->assertNotNull($id20);
+        $this->assertSame(20, $id20->id);
+
+        // with DB
+        AuthFake::databaseSet([
+            ['id' => 1, 'name' => 'John'],
+            ['id' => 2, 'name' => 'Jane'],
+        ]);
+
+        $id1 = $this->getAuthFake()->fromId(1);
+        $this->assertNotNull($id1);
+        $this->assertSame('John', $id1->name);
+        $this->assertSame(1, $id1->id);
+
+        $id3 = $this->getAuthFake()->fromId(3);
+        $this->assertNull($id3);
+    }
+
+    public function testFromEmail(): void
+    {
+        $this->enable();
+        $email20 = $this->getAuthFake()->fromEmail('20@test.com');
+        $this->assertNotNull($email20);
+        $this->assertSame('20@test.com', $email20->email);
+
+        // with DB
+        AuthFake::databaseSet([
+            ['id' => 1, 'name' => 'John', 'email' => 'john@test.com'],
+            ['id' => 2, 'name' => 'Jane', 'email' => 'jane@test.com']
+        ]);
+
+        $email1 = $this->getAuthFake()->fromEmail('john@test.com');
+        $this->assertNotNull($email1);
+        $this->assertSame('John', $email1->name);
+        $this->assertSame('john@test.com', $email1->email);
+
+        $email3 = $this->getAuthFake()->fromEmail('supun@test.com');
+        $this->assertNull($email3);
+    }
+
+    public function testFromUsername(): void
+    {
+        $this->enable();
+        $username20 = $this->getAuthFake()->fromUsername('user20');
+        $this->assertNotNull($username20);
+        $this->assertSame('user20', $username20->username);
+
+        // with DB
+        AuthFake::databaseSet([
+            ['id' => 1, 'name' => 'John', 'username' => 'john'],
+            ['id' => 2, 'name' => 'Jane', 'username' => 'jane']
+        ]);
+
+        $username1 = $this->getAuthFake()->fromUsername('john');
+        $this->assertNotNull($username1);
+        $this->assertSame('John', $username1->name);
+        $this->assertSame('john', $username1->username);
+
+        $username3 = $this->getAuthFake()->fromUsername('supun');
+        $this->assertNull($username3);
+    }
+
+    public function testFromIds(): void
+    {
+        $this->enable();
+        $ids = $this->getAuthFake()->fromIds([1, 2, 3]);
+        $this->assertCount(3, $ids);
+        $this->assertSame(1, $ids[1]->id);
+        $this->assertSame(2, $ids[2]->id);
+        $this->assertSame(3, $ids[3]->id);
+
+        // with DB
+        AuthFake::databaseSet([
+            ['id' => 1, 'name' => 'John', 'username' => 'john'],
+            ['id' => 2, 'name' => 'Jane', 'username' => 'jane']
+        ]);
+
+        $ids = $this->getAuthFake()->fromIds([1, 2, 3]);
+        $this->assertCount(2, $ids);
+        $this->assertSame(1, $ids[1]->id);
+        $this->assertSame(2, $ids[2]->id);
+    }
+
+    public function testFromUsernames(): void
+    {
+        $this->enable();
+        $usernames = $this->getAuthFake()->fromUsernames(['user1', 'user2', 'user3']);
+        $this->assertCount(3, $usernames);
+        $this->assertSame('user1', $usernames['user1']->username);
+        $this->assertSame('user2', $usernames['user2']->username);
+        $this->assertSame('user3', $usernames['user3']->username);
+
+        // with DB
+        AuthFake::databaseSet([
+            ['id' => 1, 'name' => 'John', 'username' => 'john'],
+            ['id' => 2, 'name' => 'Jane', 'username' => 'jane']
+        ]);
+
+        $usernames = $this->getAuthFake()->fromUsernames(['john', 'jane', 'supun']);
+        $this->assertCount(2, $usernames);
+        $this->assertSame('john', $usernames['john']->username);
+        $this->assertSame('jane', $usernames['jane']->username);
+    }
+
+    public function testFromEmails(): void
+    {
+        $this->enable();
+        $emails = $this->getAuthFake()->fromEmails(['user1@test.com', 'user2@test.com']);
+        $this->assertCount(2, $emails);
+        $this->assertSame('user1@test.com', $emails['user1@test.com']->email);
+        $this->assertSame('user2@test.com', $emails['user2@test.com']->email);
+
+        // with DB
+        AuthFake::databaseSet([
+            ['id' => 1, 'name' => 'John', 'email' => 'john@test.com'],
+            ['id' => 2, 'name' => 'Jane', 'email' => 'jane@test.com']
+        ]);
+
+        $emails = $this->getAuthFake()->fromEmails(['john@test.com', 'jane@test.com', 'roger@test.com']);
+        $this->assertCount(2, $emails);
+        $this->assertSame('john@test.com', $emails['john@test.com']->email);
+        $this->assertSame('jane@test.com', $emails['jane@test.com']->email);
+    }
+
+}