+++ tests/Unit/Auth/AuthSymfonyTest.php
@@ -7,9 +7,4 @@ use Hyvor\Internal\Tests\SymfonyTestCase;
 class AuthSymfonyTest extends SymfonyTestCase
 {
     use AuthTestTrait;
-
-    protected function getContainer(): \Symfony\Component\DependencyInjection\Container
-    {
-        return $this->container;
-    }
 }